fix: moved util to logic
